% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btc}
\alias{btc}
\title{Bitcoin daily close price and open interest data}
\format{
A data frame with 503 rows and 15 variables.
}
\source{
\url{https://cryptoquant.com}
}
\usage{
btc
}
\description{
A dataset of daily close values for bitcoin prices and open interests
in bitcoin futures between 2019-12-10 and 2021-04-25. The data was obtained
from CryptoQuant. The dataset also has some additional fields that are
results of a simulation of trading with a simple algorithm on an exchange.
}
\keyword{datasets}
