% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_collapse_rows.R
\name{kable_collapse_rows}
\alias{kable_collapse_rows}
\title{Collapse repeated rows to multirow cell.}
\usage{
kable_collapse_rows(tbl_code, cols_idx = 1:2)
}
\arguments{
\item{tbl_code}{markdown/tex code that describes the table to be displayed.
For example, output object from kable_default(), kable_long(), kable_wide()
are good candidates.}

\item{cols_idx}{integer vector, indices of columns to be put into multi-row
cells. Default is first 2 columns.}
}
\value{
markdown/tex code block for a table that can be rendered inside a R
markdown file when outputting a pdf document.
}
\description{
Collapse same values in columns into multirow cells. You may want to do this
when the table is too wide. After doing this, the readability of the table is
often improved.
}
\examples{
inst/examples/ex-kable_collapse_rows.R
}
