% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_default.R
\name{kable_default}
\alias{kable_default}
\title{Display table in default mode.}
\usage{
kable_default(tbl, ...)
}
\arguments{
\item{tbl}{Data frame, the table to be displayed.}

\item{...}{Other arguments for \code{kntr::kable()}, for example, digits = 2,
caption = 'Table Caption', and etc.}
}
\value{
markdown/tex code block for a table that can be render inside a R
markdown file when outputting a pdf document.
}
\description{
This function displays the given table at the exact position on the pdf page
where the table is inserted. It calls \code{knitr::kable()} under the hood, with
\code{booktabs=TRUE} and \code{format='latex'}.
}
\examples{
inst/examples/ex-kable_default.R
}
