% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_landscape.R
\name{to_landscape}
\alias{to_landscape}
\title{Change the table display to landscape.}
\usage{
to_landscape(tbl_code, margin = "1cm")
}
\arguments{
\item{tbl_code}{markdown/tex code that describes the table to be displayed.
For example, output object from kable_default(), kable_long(), kable_wide()
are good candidates.}
}
\value{
markdown/tex code block for a table that can be rendered inside a R
markdown file when outputting a pdf document.
}
\description{
After applying \code{kable_long()} or \code{kable_wide()} to a table, the resulting
display sometimes can still fail to fit on one page. When that happens, you
can apply this function to change the table display to landscape.
}
\examples{
inst/examples/ex-to_landscape.R
}
