% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_wide.R
\name{kable_wide}
\alias{kable_wide}
\title{Display wide tables.}
\usage{
kable_wide(tbl, font_size = "auto", ...)
}
\arguments{
\item{tbl}{Data frame, the table to be displayed.}

\item{font_size}{Number, font size of the table content. Default = 'auto',
which will automatically pick a font size to fit the table on the page.}

\item{...}{Other arguments for \code{kntr::kable()}, for example, digits = 2,
caption = 'Table Caption', and etc.}
}
\value{
markdown/tex code block for a table that can be rendered inside a R
markdown file when outputting a pdf document.
}
\description{
This function works great if you need to fit a wide table on one page. It
builds on top of \code{kable_default()}.
}
\examples{
inst/examples/ex-kable_wide.R
}
